changes to nav.py window management
* moved minimise and maximise commands from _caster
* added workspace commands
* expanded window snapping functionality
