+++ caster/apps/dragon.py
@@ -36,6 +36,9 @@ class DragonRule(MergeRule):
         'normal mode':
             R(Playback([(["normal", "mode", "on"], 0.0)]),
               rdescript="Dragon: Normal Mode"),
+        'command mode':
+            R(Playback([(["command", "mode", "on"], 0.0)]),
+              rdescript="Dragon: Normal Mode"),
         'com on':
             R(Playback([(["command", "mode", "on"], 0.0)]),
               rdescript="Dragon: Command Mode (On)"),
